brancher MySQL work in progress

+++ brancher/base.py
@@ -1,11 +1,22 @@
+from StringIO import StringIO
+
 from django.conf import settings
 from django.db import connection
 from subprocess import check_output
+
+from model_utils import Choices
+
+# noinspection PyInterpreter
 
 
 class DbNameMixin(object):
     _branch_name = None
     _db_name = None
+
+    DB_ENGINES = Choices(
+        'postgresql',
+        'mysql'
+    )
 
     def add_arguments(self, parser):
         parser.add_argument('--branch_name', help="This overrides the git branch as the branch name")
@@ -31,8 +42,54 @@ class DbNameMixin(object):
 
     def create_database(self):
         with connection.cursor() as cursor:
-            cursor.execute("CREATE DATABASE {branch} template {name};".format(branch=self.full_branched_db_name, name=self.get_db_name))
+            if self.is_postgresql:
+                cursor.execute("CREATE DATABASE {branch} template {name};".format(branch=self.full_branched_db_name, name=self.get_db_name))
+            elif self.is_mysql:
+                data = StringIO()
+                data.write(cursor.execute())
+                # dump 'default' database
+                    with open(filename, 'w') as f:
+                        cmd = ['mysqldump',
+                               '--user={}'.format(db_user),
+                               '--password={}'.format(db_pass),
+                               '{}'.format(DATABASES['default']['NAME']+'__master')]
+                        f.write(subprocess.check_output(cmd))
 
-    def drop_dataase(self):
+                # create new database
+                cmd = ['mysql',
+                       '--user={}'.format(db_user),
+                       '--password={}'.format(db_pass),
+                       '--execute=CREATE DATABASE {};'.format(db_name)]
+                subprocess.check_call(cmd)
+
+                # mysql import dump.sql to new db
+                with open(filename, 'r') as f:
+                    cmd = ['mysql',
+                           '--user={}'.format(db_user),
+                           '--password={}'.format(db_pass),
+                           db_name]
+                    subprocess.check_call(cmd, stdin=f)
+
+    def drop_database(self):
         with connection.cursor() as cursor:
-            cursor.execute("DROP DATABASE IF EXISTS {database};".format(database=self.full_branched_db_name))
+            if self.is_postgresql:
+                cursor.execute("DROP DATABASE IF EXISTS {database};".format(database=self.full_branched_db_name))
+            elif self.is_mysql:
+                raise NotImplementedError()
+
+    @property
+    def db_engine(self):
+        if settings.DATABASES['default']['ENGINE'] == 'django.db.backends.mysql':
+            return self.DB_ENGINES.mysql
+        elif settings.DATABASES['default']['ENGINE'] == 'django.db.backends.postgresql':
+            return self.DB_ENGINES.postgresql
+        else:
+            raise Exception('Not Implemented')
+
+    @property
+    def is_postgresql(self):
+        return self.db_engine == self.DB_ENGINES.postgresql
+
+    @property
+    def is_mysql(self):
+        return self.db_engine == self.DB_ENGINES.mysql
